fix: cast IMPORT_STUCK_THRESHOLD_MINUTES to int in config
env() returns a string when the .env key is actually present but the
literal int fallback only when it's absent. That inconsistency reached
Carbon's addMinutes(), which requires int|float and throws on a numeric
string — surfaced once .env.example (and therefore any fresh .env) started
carrying IMPORT_STUCK_THRESHOLD_MINUTES.

+++ config/imports.php
@@ -5,7 +5,7 @@ return [
     // considered stuck. Drives both the /admin/sellers UI banner and the
     // monitor job's email alert, and doubles as the monitor job's
     // re-check interval — one knob, not two, on purpose.
-    'stuck_after_minutes' => env('IMPORT_STUCK_THRESHOLD_MINUTES', 15),
+    'stuck_after_minutes' => (int) env('IMPORT_STUCK_THRESHOLD_MINUTES', 15),
 
     'notification_email' => env('IMPORT_NOTIFICATION_EMAIL', env('RFQ_NOTIFICATION_EMAIL', 'sales@example.com')),
 ];
